Updates after CR: - Created generic apilayer class - Updated endpoints names - Reduced serchable user fields in get user endpoint
