--- app/users/user_manager_exceptions.py
@@ -1,18 +1,0 @@
-class UserManagerException(Exception):
-    pass
-
-
-class NotFoundException(UserManagerException):
-    pass
-
-
-class ConflictException(UserManagerException):
-    pass
-
-
-class ServiceUnavailableException(UserManagerException):
-    pass
-
-
-class BadRequestException(UserManagerException):
-    pass
