+++ app/utils/api_layer_exceptions.py
@@ -1,0 +1,18 @@
+class BaseApiException(Exception):
+    pass
+
+
+class NotFoundException(BaseApiException):
+    pass
+
+
+class ConflictException(BaseApiException):
+    pass
+
+
+class ServiceUnavailableException(BaseApiException):
+    pass
+
+
+class BadRequestException(BaseApiException):
+    pass
